Add files via upload
Completely reworked the code. Now contains an abstract tree class and two concrete trees (A Cayley tree and a Lieb-Cayley tree). Added a variety of functionalities, including a noise generator to simulate local disorder and a low-ram mode which allows to simulate spectras for large trees (M > 100) using effective diagonalization.